+++ test/core/interface.jl
@@ -1,4 +1,3 @@
-# TODO add tests for user defined types
 g = G(10)
 h = DG(10)
 @test edge(g, 1, 2) != nothing
@@ -10,3 +9,19 @@ h = DG(10)
 @test graph(h) != nothing
 @test digraph(g) != nothing
 @test digraph(h) != nothing
+
+type TestGraph <: AGraph; end
+type TestDiGraph <: ADiGraph; end
+g = TestGraph()
+h = TestDiGraph()
+
+@test nv(g) == nothing
+@test nv(h) == nothing
+@test neighbors(g, 1) == nothing
+@test edge(g, 1, 2) == nothing
+@test edge(h, 1, 2) == nothing
+@test neighbors(g, 1) == nothing
+@test in_neighbors(h, 1) == nothing
+@test out_neighbors(h, 1) == nothing
+@test graphtype(h) == nothing
+@test digraphtype(g) == nothing